Move test into optimization disabled and remove sr8426.swift
